Fixed rendering of meshes in voxel sizes != 1.0
We used the openvdb voxel size to render a mesh into a voxel field and combined it with our existing voxel field, which always has openvdb voxel size of 1.0. This resulted in the voxel field having artifacts in the signed distance values. This was particularly apparent when using small voxel sizes.

This solution, while it works, incurs a small performance hit, since a duplicated mesh with transformed values is created. This is probably not dramatic, but we should fix this in version 2.0 by using the openvdb voxel size and eliminate our extern world to voxel calculations. Fixes issue #8

Also removed an unused function.