Fixed lattice voxelization generated wrong SDF
The latice voxelization function did not convert the signed distance value to voxel coordinates. So for voxel sizes deviating from 1, the function produced corrupt voxel fields. This was especially visible for very small voxel sizes. In. voxel sizes around the typical ones we use at LEAP 71 - 0.1 .. 1.0, this was not very visible. Fixes issue #9